==> fl-politics/elections/2020/general/Results/Why did Florida go red?/1_download.ipynb <==
# Download data

We'll need to download the data to analyze first. We will need the 2020 election results,as well as the 2016 results, which need to be manually downloaded from here: https://results.elections.myflorida.com/downloadresults.asp?ElectionDate=11/8/2016&DATAMODE=. 

import requests

## Get county level results

county_results_url = 'https://flelectionfiles.floridados.gov/enightfilespublic/20201103_ElecResultsFL.txt'
county_results_response = requests.get(county_results_url)
with open('data/2020_county_results.txt', 'w') as txtfile:
    txtfile.write(county_results_response.text)
    

## Get county turnout data

county_turnout_url = 'https://dos.myflorida.com/media/703608/1-party-by-county.xlsx'
county_turnout_response = requests.get(county_turnout_url)
with open('data/2020_county_turnout.xls', 'wb') as datafile:
    datafile.write(county_turnout_response.content)
