==> fl-politics/elections/2020/general/Results/Why did Florida go red?/2_county_margins_analysis.ipynb <==
# Florida went red. Where did Trump and Biden gain and lose?

An analysis of Florida voting data to determine what helped make the state go for Donald Trump. We'll look at county voting data and compare changes in margins from 2016 to 2020.

import pandas as pd
import numpy as np

## How did counties vote?
We'll analyze the county results from the state DOE. Find those here: https://floridaelectionwatch.gov/Downloads.

county_data = pd.read_csv('data/2020_county_results.txt', delimiter='\t')
county_data.head()

# Filter for the presidential race only.
county_pres_data = county_data[county_data["RaceName"] == "President of the United States"]
county_pres_data.head()

#### Reshape the data so counites have columns for each candidate.

# Pivot candidate rows to columns, preserving the vote counts as values.
fl_county_results = county_pres_data.pivot(index='CountyName', columns='CanNameFirst', values=['CanVotes'])

# Flatten multi-index dataframe
fl_county_results.columns = fl_county_results.columns.get_level_values(1)

# Reset index and make candidate names the headers
fl_county_results.columns = pd.Series(fl_county_results.columns.tolist()).apply(pd.Series).sum(axis=1)
fl_county_results = fl_county_results.reset_index()

fl_county_results.head()

#### Get total votes for each county.

fl_county_results['total_votes'] = fl_county_results.iloc[:, 1:].sum(axis=1)
fl_county_results.head()

# Make a slice of this dataset to narrow down Trump vs. Biden results.
biden_vs_trump = fl_county_results[['CountyName', 'Biden', 'Trump', 'total_votes']]

### Compare data with 2016
Load in results from 2016 and do the same reshaping so we can compare the two elections. Data from 2016 can be manually downloaded here: https://results.elections.myflorida.com/downloadresults.asp?ElectionDate=11/8/2016&DATAMODE=

results_2016 = pd.read_csv('data/11082016Election.txt', delimiter='\t')
results_2016.head(2)

# Reshape using same method as 2020 data, but change relevant fields to match 2016 data format.
county_pres_data_2016 = results_2016[results_2016["OfficeDesc"] == "President of the United States"]

pres_data_2016 = county_pres_data_2016.pivot(index='CountyName', columns='CanNameLast', values=['CanVotes'])
pres_data_2016.columns = pres_data_2016.columns.get_level_values(1)

pres_data_2016.columns = pd.Series(pres_data_2016.columns.tolist()).apply(pd.Series).sum(axis=1)
pres_data_2016 = pres_data_2016.reset_index()
pres_data_2016['total_votes'] = pres_data_2016.iloc[:, 1:].sum(axis=1)

pres_data_2016.head()

# Narrow to Clinton vs. Trump data
clinton_vs_trump = pres_data_2016[['CountyName', 'Clinton', 'Trump', 'total_votes']]

### Merge 2020 and 2016 results

results_2020_vs_2016 = biden_vs_trump.merge(clinton_vs_trump, on='CountyName', suffixes=('_2020', '_2016'))
results_2020_vs_2016.head()

#### Calculate vote percentages.

results_2020_vs_2016['Biden_pct'] = results_2020_vs_2016['Biden'] / results_2020_vs_2016['total_votes_2020'] * 100
results_2020_vs_2016['Trump_2020_pct'] = results_2020_vs_2016['Trump_2020'] / results_2020_vs_2016['total_votes_2020'] * 100
results_2020_vs_2016['Clinton_pct'] = results_2020_vs_2016['Clinton'] / results_2020_vs_2016['total_votes_2016'] * 100
results_2020_vs_2016['Trump_2016_pct'] = results_2020_vs_2016['Trump_2016'] / results_2020_vs_2016['total_votes_2016'] * 100
results_2020_vs_2016.head()

#### Flag who won and if a county flipped.

results_2020_vs_2016['2020_winner'] = results_2020_vs_2016.apply(lambda x: 'Trump' if x['Trump_2020_pct'] > x['Biden_pct'] else 'Biden', axis=1)
results_2020_vs_2016['2016_winner'] = results_2020_vs_2016.apply(lambda x: 'Trump' if x['Trump_2016_pct'] > x['Clinton_pct'] else 'Clinton', axis=1)


def check_flips(row):
    if row['2020_winner'] == row['2016_winner']:
        return ''
    elif row['2020_winner'] == 'Biden' and row['2016_winner'] == 'Clinton':
        return ''
    elif row['2020_winner'] == 'Biden' and row['2016_winner'] == 'Trump':
        return 'Biden'
    elif row['2020_winner'] == 'Trump' and row['2016_winner'] == 'Clinton':
        return 'Trump'
    
    
results_2020_vs_2016['flip'] = results_2020_vs_2016.apply(check_flips, axis=1)
results_2020_vs_2016.head()

### Which counties flipped?
Three Florida counties flipped to Biden, according to the data. Trump won Duval, Pinellas and Seminole in 2016, but not in 2020.

results_2020_vs_2016[results_2020_vs_2016['flip'] != '']

## How did county margins change?

margins_data = results_2020_vs_2016
margins_data['Dem_change'] = margins_data['Biden_pct'] - margins_data['Clinton_pct']
margins_data['GOP_change'] = margins_data['Trump_2020_pct'] - margins_data['Trump_2016_pct']

margins_data.to_csv('data/vote_margin_changes.csv', index=False)
margins_data.head()

### Where did Biden see the most changes?

#### Biden top margin losses

Biden saw a big loss in Miami-Dade County, as has been reported. He also lost traction in and Osceola, Hendry, Glades and Baker counties compared to Clinton in 2016.

margins_data.sort_values('Dem_change', ascending=True).head()

#### Biden top margin gains

Biden managed to make minor improvements to the Democratic vote share compared to Clinton in Okaloosa, Santa Rosa, Clay, St. Johns and Escambia counties.

margins_data.sort_values('Dem_change', ascending=False).head()

#### Count how many counties Biden saw gains and losses in.

biden_gains_count = margins_data[margins_data["Dem_change"] > 0]
biden_loss_count = margins_data[margins_data["Dem_change"] < 0]
print("Gains:", biden_gains_count.shape[0])
print("Losses:", biden_loss_count.shape[0])

### Where did Trump see the most changes?

#### Trump top margin losses

Compared to Biden, Trump saw much less of a decline in support in counties. The biggest loss is in Okaloosa, where he won 2.7 points less this election than he did in 2016.

margins_data.sort_values('GOP_change', ascending=True).head()

#### Trump top margin gains

Trump's margin gains also outweigh the improvements Biden made. Trump managed to improve his vote share gain by nearly 12 points in Miami-Dade counties, while also upping his margins in Osceola and Hendry counties.

margins_data.sort_values('GOP_change', ascending=False).head()

#### Count how many counties Trump saw gains and losses in.

trump_gains_count = margins_data[margins_data['GOP_change'] > 0]
trump_loss_count = margins_data[margins_data['GOP_change'] < 0]
print('Gains:', trump_gains_count.shape[0])
print('Losses:', trump_loss_count.shape[0])

### Margins of victory
Now, compare overall margin of victory changes by calculating the point differences between the candidates' vote percentages.

margins_data['2016_mov'] = abs(margins_data['Clinton_pct'] - margins_data['Trump_2016_pct'])
margins_data['2020_mov'] = abs(margins_data['Biden_pct'] - margins_data['Trump_2020_pct'])
margins_data['mov_change'] = margins_data['2020_mov'] - margins_data['2016_mov']

# Save data
margins_data.to_csv('data/mov_changes.csv', index=False)

#### Trump's margin of victory gains
Trump improved his margin of victory most in Hendry County followed by Glades, Calhoun, Baker and Okeechobee counties.

trump_wins = margins_data[margins_data['2020_winner'] == 'Trump']
trump_wins.sort_values('mov_change', ascending=False).head()

#### Trump's margin of victory losses
Trump's margin of victory shrunk most in Okaloosa County, followed by Clay, Santa Rosa, St. Johns and Escambia counties.

trump_wins.sort_values('mov_change', ascending=True).head()

#### Count how many counties Trump had gains and losses of victory margins.

trump_mov_gains_count = trump_wins[trump_wins['mov_change'] > 0]
trump_mov_loss_count = trump_wins[trump_wins['mov_change'] < 0]
print('Gains:', trump_mov_gains_count.shape[0])
print('Losses:', trump_mov_loss_count.shape[0])

#### Bidens's margin of victory gains
Biden upped his margin of victory compared to Clinton most in Alachua County, followed by Leon, Duval, Seminole and Hillsborough counties.

biden_wins = margins_data[(margins_data['2020_winner'] == 'Biden') & (margins_data['flip'] == '')]
biden_wins.sort_values('mov_change', ascending=False).head()

#### Bidens's margin of victory losses
Biden took big losses to his margin of victory, compared to Clinton, in Miami-Dade County, followed by Osceola, Broward, Palm Beach and Orange counties.

biden_wins.sort_values('mov_change', ascending=True).head()

biden_mov_gains_count = biden_wins[biden_wins['mov_change'] > 0]
biden_mov_loss_count = biden_wins[biden_wins['mov_change'] < 0]
print('Gains:', biden_mov_gains_count.shape[0])
print('Losses:', biden_mov_loss_count.shape[0])

## Statewide margins

Trump won Florida in 2016 and 2020. Let's take a look at how his margins of victory compare.

statewide_results = results_2020_vs_2016[['Biden','Trump_2020','total_votes_2020','Clinton','Trump_2016','total_votes_2016']]
statewide_results = statewide_results.agg(['sum'])

statewide_results.head()

statewide_results['2020_Trump_pct'] = statewide_results['Trump_2020'] / statewide_results['total_votes_2020'] * 100
statewide_results['2020_Biden_pct'] = statewide_results['Biden'] / statewide_results['total_votes_2020'] * 100
statewide_results['2016_Clinton_pct'] = statewide_results['Clinton'] / statewide_results['total_votes_2016'] * 100
statewide_results['2016_Trump_pct'] = statewide_results['Trump_2016'] / statewide_results['total_votes_2016'] * 100
statewide_results['2016_Trump_pct'][0]

print("Trump's 2016 margin of victory:", statewide_results['2016_Trump_pct'][0] - statewide_results['2016_Clinton_pct'][0])
print("Trump's 2020 margin of victory:", statewide_results['2020_Trump_pct'][0] - statewide_results['2020_Biden_pct'][0])
