==> fl-politics/elections/2020/general/Results/Why did Florida go red?/3_county_votes_analysis.ipynb <==
# Where did Trump and Biden gain and lose votes?
Trump won Florida and tripled his margin of victory, but this analysis will look at where he and Biden gained and lost the votes that led to that.

import pandas as pd
import numpy as np

### Load a slice of the data produced in the previous notebook.

votes_df = pd.read_csv('data/vote_margin_changes.csv')
votes_df = votes_df[['CountyName', 'Biden', 'Trump_2020', 'total_votes_2020', 'Clinton', 'Trump_2016', 'total_votes_2016']]
votes_df.head()

### Calculate the difference in the raw vote totals for 2016 and 2020

votes_df['d_vote_change'] = votes_df['Biden'] - votes_df['Clinton']
votes_df['r_vote_change'] = votes_df['Trump_2020'] - votes_df['Trump_2016']

### How much did Trump and Biden gain in each county?

votes_df['d_vote_pct_change'] = votes_df['d_vote_change'] / votes_df['Clinton'] * 100
votes_df['r_vote_pct_change'] = votes_df['r_vote_change'] / votes_df['Trump_2016'] * 100

votes_df.sort_values('r_vote_pct_change', ascending=False).head(10)

votes_df.sort_values('d_vote_pct_change', ascending=False).head(10)

### Compare differences to see how each performed in out-collecting the other for votes
For Biden to have overtaken Trump in Florida, he would have had to exceed Trump's increase in new votes, vote-for-vote compared to 2016. 

We'll subtract Trump's gains from Biden to give us a net difference. Positive values mean Biden outperformed Trump, while negative values mean Trump outperformed Biden.

votes_df['net_vote_change'] = votes_df['d_vote_change'] - votes_df['r_vote_change']
votes_df.sort_values('net_vote_change', ascending=True)

votes_df.drop(['total_votes_2020', 'total_votes_2016'], axis=1)

### How many counties did each perform better and worse? 

biden_gains = votes_df[votes_df["net_vote_change"] > 0]
trump_gains = votes_df[votes_df["net_vote_change"] < 0]

print("Biden did better in {} counties".format(biden_gains.shape[0]))
print("Trump did better in {} counties".format(trump_gains.shape[0]))

### How does this translate to the number of votes?

biden_vote_gain = biden_gains.net_vote_change.sum()
trump_vote_gain = abs(trump_gains.net_vote_change).sum()

trump_vote_gain + biden_vote_gain
# print("Biden did better in {} counties".format(biden_vote_gain))
# print("Trump did better in {} counties".format(trump_vote_gain))
